+++ src/.fpm_spine_install.rb
@@ -39,9 +39,8 @@ if target_repo == "1"
   system("curl #{core}/#{target}.tar.gz -o #{home}/.fpm/install/#{target}.tar.gz")
   puts "Verifying package intergrity..."
   system("sha256sum #{home}/.fpm/install/#{target}.tar.gz > /tmp/fpm_checksum")
-  system("curl -fs #{core}/#{target}.tar.gz -o /tmp/pkg.tar.gz")
-  system("sha256sum /tmp/pkg.tar.gz > /tmp/fpm_compare_sum")
-  solid = FileUtils.identical?("/tmp/fpm_checksum", "/tmp/fpm_compare_sum")
+  system("curl -fs #{core}/index/#{target}.hash -o /tmp/#{target}.hash")
+  solid = FileUtils.identical?("/tmp/fpm_checksum", "/tmp/#{target}.hash")
   if solid == "false"
     puts "Unable to verify package integrity."
     puts "Install anyways? [y/N]"
@@ -53,7 +52,7 @@ if target_repo == "1"
       puts "Cancelling..."
       puts "Cleanup directories..."
       system("rm -f #{home}/.fpm/install/*")
-      system("rm -f /tmp/fpm_*")
+      system("rm -f /tmp/fpm_* /tmp/*.hash")
       abort "fpm: fatal: package integrity could not be verified."
     end
   else
@@ -62,19 +61,19 @@ if target_repo == "1"
   puts "try unpack".blue
   system("tar -xzvf #{home}/.fpm/install/#{target}.tar.gz -C #{home}/.fpm/install")
   puts "If there was no errors, the package should have been unpacked.".yellow
-  puts "try transfer".blue
-  sleep(1)
-  system("cp #{home}/.fpm/install/#{target} #{home}/.fpm/pkgs/#{target}")
+  puts "run construct script".blue
+  system("ruby #{home}/.fpm/install/#{target}.construct.rb")
   puts "If all went well, your package should be installed. If you haven't already, add #{home}/.fpm/pkgs to your PATH variable.".green
 elsif target_repo == "2"
+  puts "Note: packages in the 'live' repository are not subject to integrity checks.".red
+  puts "This is because the code is often updated frequently and isn't intended to be mainstream."
   puts "Trying to get #{target} from live.".yellow
   system("curl #{live}/#{target}.tar.gz -o #{home}/.fpm/install/#{target}.tar.gz")
   puts "try unpack".blue
   system("tar -xzvf #{home}/.fpm/install/#{target}.tar.gz #{home}/.fpm/install")
   puts "If there was no errors, the package should have been unpacked.".yellow
-  puts "try transfer".blue
-  eval(File.read("fpkgd"), binding)
-  system("cp #{home}/.fpm/install/#{target} #{home}/.fpm/pkgs/#{target}")
+  puts "run construct script".blue
+  system("ruby #{home}/.fpm/install/#{target}.construct.rb")
   puts "If all went well, your package should be installed. If you haven't already, add #{home}/.fpm/pkgs to your PATH variable.".green
 else
   puts "unknown repo. exiting."