+++ src/.installation/fpm2_install_head.cpp
@@ -7,39 +7,36 @@
 #include <string.h>
 
 using namespace std;
+void installme() {
+  string home = getenv("HOME");
+  string createdirs = "mkdir " + home + "/.fpm && mkdir " + home + "/.fpm/install && mkdir " + home + "/.fpm/pkgs && mkdir " + home + "/.fpm/pkgs/help";
+  cout << "\nBeginning installation...";
+  cout << "\nInstalling ruby dependencies...";
+  system("gem install colorize");
+  cout << "\nDependency 'colorize' installed.";
+  system("gem install fileutils");
+  cout << "\nDependency 'fileutils' installed.";
+  cout << "\nGetting remote files...";
+  system("ruby install_nix.rb");
+  cout << "\nCreating directories...";
+  system(createdirs.c_str());
+  cout << "\nDone!";
+}
 // code starts here
 int main() {
-  /* commands - get home dir */ string home = getenv("HOME");
-  const string wininstall = "ruby install_win.rb";
-  const string nixinstall = "ruby install_nix.rb";
-  const string installdep = "gem install colorize";
-  const string installdep2 = "gem install fileutils";
-  string createdirs = "mkdir " + home + "/.fpm && mkdir " + home + "/.fpm/install && mkdir " + home + "/.fpm/pkgs && mkdir " + home + "/.fpm/pkgs/help"
-  // install function
-  function installme() {
-    cout << "\nBeginning installation...";
-    cout << "\nInstalling ruby dependencies...";
-    system(installdep.c_str());
-    cout << "\nDependency 'colorize' installed.";
-    system(installdep2.c_str());
-    cout << "\nDependency 'fileutils' installed.";
-    cout << "\nGetting remote files...";
-    system(nixinstall.c_str());
-    cout << "\nCreating directories...";
-    system(createdirs.c_str());
-    cout << "\nDone!"
-  }
-
-  // variables
-  string input;
+  /* commands - get home dir */
+  int input;
   // ask for operating system
   cout << "\nReady to install? \n[1.] Yes\n2. No";
   cout << "\nYour response: ";
   cin >> input;
   switch (input) {
     case 1:
-      installme
+      installme();
       break;
     case 2:
+      cout << "\nOkay, cancelling...";
+      return 1;
+      break;
   }
 }